// Week6/homework/src/pages/movies/detail/model/movieDetail.ts
export type MovieDetail = {
  id: number
  title: string
  originalTitle: string
  originalLanguage: string
  releaseDate: string
  status: string
  runtimeMinutes: number
  voteAverage: number
  voteCount: number
  overview: string
  posterUrl: string
  backdropUrl: string
  genres: string[]
  productionCountries: string[]
  spokenLanguages: string[]
  budgetUsd: number
  revenueUsd: number
}